[CARBONDATA-2224][File Level Reader Support] Refactoring of #2055
Review comment fixes and refactoring of #2055

This closes #2069
